Seperated the transactoins fetching from the useGetItems hook
