+++ src/hooks/index.js
@@ -1,5 +1,6 @@
 import useAddItem from "./useAddItem.js";
 import useGetItems from "./useGetItems.js";
 import useTotalBalanceStore from "./useTotalBalanceStore.js";
+import fetchTransactions from "./fetchTransactions.js";
 
-export { useAddItem, useGetItems, useTotalBalanceStore };
+export { useAddItem, useGetItems, useTotalBalanceStore, fetchTransactions };
