+++ src/hooks/useGetItems.js
@@ -1,26 +1,11 @@
 import { useState, useEffect } from "react";
-import {
-  query,
-  collection,
-  where,
-  orderBy,
-  limit,
-  getDocs,
-  startAfter,
-  limitToLast,
-  endAt,
-} from "firebase/firestore";
-import { db } from "../firebase/firebase.js";
 import { useUser } from "@clerk/clerk-react";
+import { fetchTransactions } from "./";
 
 function useGetItems(itemsPerPage = 11) {
   const [transactions, setTransactions] = useState([]);
 
-  const defaultSort = {
-    column: "datetime",
-    type: "desc",
-  };
-  const [sortingQuery, setSortingQuery] = useState(defaultSort);
+  const [sortingQuery, setSortingQuery] = useState(null);
   const [pagination, setpagination] = useState({
     currentPage: 1,
     direction: "next",
@@ -29,25 +14,16 @@ function useGetItems(itemsPerPage = 11) {
     isLast: false,
   });
 
-  const transactionsRef = collection(db, "transactions");
   const { user } = useUser();
 
   const getItems = async () => {
-    const directionQuery =
-      pagination.firstVisible && pagination.currentPage !== 1
-        ? pagination.direction == "next"
-          ? [startAfter(pagination.lastVisible), limit(itemsPerPage)]
-          : [endAt(pagination.firstVisible), limitToLast(itemsPerPage)]
-        : [limit(itemsPerPage)];
-
     try {
-      const myQuery = query(
-        transactionsRef,
-        where("uid", "==", user.id),
-        orderBy(sortingQuery.column, sortingQuery.type),
-        ...directionQuery
-      );
-      const snapshot = await getDocs(myQuery);
+      const snapshot = await fetchTransactions({
+        uid: user.id,
+        sortingQuery: sortingQuery,
+        pagination: pagination,
+        itemsPerPage: itemsPerPage,
+      });
 
       let cleanedData = [];
 
@@ -75,7 +51,7 @@ function useGetItems(itemsPerPage = 11) {
     });
 
     if (sortQuery) {
-      setSortingQuery(sortQuery.type ? sortQuery : defaultSort);
+      setSortingQuery(sortQuery.type ? sortQuery : null);
     }
   };
